+++ src/main/java/kr/co/mz/sns/config/security/CustomUserDetailService.java
@@ -1,7 +1,7 @@
 package kr.co.mz.sns.config.security;
 
 import java.util.List;
-import kr.co.mz.sns.dto.UserDto;
+import kr.co.mz.sns.dto.user.UserDto;
 import kr.co.mz.sns.repository.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,20 +13,20 @@ import org.springframework.stereotype.Service;
 @Service
 public class CustomUserDetailService implements UserDetailsService {
 
-    private final UserRepository userRepository;
-    private final ModelMapper modelMapper;
+  private final UserRepository userRepository;
+  private final ModelMapper modelMapper;
 
-    @Autowired
-    public CustomUserDetailService(UserRepository userRepository, ModelMapper modelMapper) {
-        this.userRepository = userRepository;
-        this.modelMapper = modelMapper;
-    }
+  @Autowired
+  public CustomUserDetailService(UserRepository userRepository, ModelMapper modelMapper) {
+    this.userRepository = userRepository;
+    this.modelMapper = modelMapper;
+  }
 
-    @Override
-    public CustomUserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        var userEntity = userRepository.findByEmail(email)
-            .orElseThrow(() -> new UsernameNotFoundException("Username not found"));
-        return new CustomUserDetails(modelMapper.map(userEntity, UserDto.class),
-            List.of(new SimpleGrantedAuthority(userEntity.getRole())));
-    }
+  @Override
+  public CustomUserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+    var userEntity = userRepository.findByEmail(email)
+        .orElseThrow(() -> new UsernameNotFoundException("Username not found"));
+    return new CustomUserDetails(modelMapper.map(userEntity, UserDto.class),
+        List.of(new SimpleGrantedAuthority(userEntity.getRole())));
+  }
 }
