--- src/main/java/kr/co/mz/sns/dto/PostLikeDto.java
@@ -1,16 +1,0 @@
-package kr.co.mz.sns.dto;
-
-import java.time.LocalDateTime;
-import lombok.Data;
-
-@Data
-public class PostLikeDto {
-
-    private Long seq;
-
-    private Long userSeq;
-
-    private Long postSeq;
-
-    private LocalDateTime createdAt;
-}
