+++ src/main/java/kr/co/mz/sns/dto/login/AuthResponseDto.java
@@ -1,4 +1,4 @@
-package kr.co.mz.sns.dto;
+package kr.co.mz.sns.dto.login;
 
 import lombok.Data;
