+++ src/main/java/kr/co/mz/sns/dto/post/PostLikeDto.java
@@ -1,0 +1,16 @@
+package kr.co.mz.sns.dto.post;
+
+import java.time.LocalDateTime;
+import lombok.Data;
+
+@Data
+public class PostLikeDto {
+
+  private Long seq;
+
+  private Long userSeq;
+
+  private Long postSeq;
+
+  private LocalDateTime createdAt;
+}
