+++ src/main/java/kr/co/mz/sns/dto/post/PostSearchDto.java
@@ -1,9 +1,9 @@
-package kr.co.mz.sns.dto;
+package kr.co.mz.sns.dto.post;
 
 import lombok.Data;
 
 @Data
 public class PostSearchDto {
 
-    String keyword;
+  String keyword;
 }
