+++ src/main/java/kr/co/mz/sns/entity/PostLikeEntity.java
@@ -6,29 +6,29 @@ import jakarta.persistence.EntityListeners;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import kr.co.mz.sns.auditing.CustomAuditingEntityListener;
+import kr.co.mz.sns.auditing.AuditorAwareImpl;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedBy;
 
 @Entity
-@EntityListeners(CustomAuditingEntityListener.class)
+@EntityListeners(AuditorAwareImpl.class)
 @Table(name = "post_like")
 @NoArgsConstructor
 @Data
 public class PostLikeEntity {
 
-    @Id
-    @Column(name = "seq")
-    private Long seq;
+  @Id
+  @Column(name = "seq")
+  private Long seq;
 
-    @CreatedBy
-    @Column(name = "user_seq")
-    private Long userSeq;
+  @CreatedBy
+  @Column(name = "user_seq")
+  private Long userSeq;
 
-    @Column(name = "post_seq")
-    private Long postSeq;
+  @Column(name = "post_seq")
+  private Long postSeq;
 
-    @Column(name = "create_at", nullable = false)
-    private LocalDateTime createdAt;
+  @Column(name = "create_at", nullable = false)
+  private LocalDateTime createdAt;
 }
