+++ src/main/java/kr/co/mz/sns/service/PostLikeService.java
@@ -1,7 +1,7 @@
 package kr.co.mz.sns.service;
 
 import java.util.List;
-import kr.co.mz.sns.dto.PostLikeDto;
+import kr.co.mz.sns.dto.post.PostLikeDto;
 import kr.co.mz.sns.entity.PostLikeEntity;
 import kr.co.mz.sns.repository.PostLikeRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,23 +15,23 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class PostLikeService {
 
-    private final PostLikeRepository postLikeRepository;
-    private final ModelMapper modelMapper;
+  private final PostLikeRepository postLikeRepository;
+  private final ModelMapper modelMapper;
 
-    @Transactional
-    public PostLikeDto insert(Long postSeq) {
-        var postLikeEntity = new PostLikeEntity();
-        postLikeEntity.setPostSeq(postSeq);
+  @Transactional
+  public PostLikeDto insert(Long postSeq) {
+    var postLikeEntity = new PostLikeEntity();
+    postLikeEntity.setPostSeq(postSeq);
 
-        return modelMapper.map(
-            postLikeRepository.save(postLikeEntity),
-            PostLikeDto.class
-        );
-    }
+    return modelMapper.map(
+        postLikeRepository.save(postLikeEntity),
+        PostLikeDto.class
+    );
+  }
 
-    public List<PostLikeDto> findAll(Long postSeq) {
-        return postLikeRepository.findByPostSeq(postSeq).stream()
-            .map(entity -> modelMapper.map(entity, PostLikeDto.class))
-            .toList();
-    }
+  public List<PostLikeDto> findAll(Long postSeq) {
+    return postLikeRepository.findByPostSeq(postSeq).stream()
+        .map(entity -> modelMapper.map(entity, PostLikeDto.class))
+        .toList();
+  }
 }
